Refactor desktop app Effect services
