+++ apps/desktop/src/app/DesktopState.ts
@@ -3,19 +3,17 @@ import * as Effect from "effect/Effect";
 import * as Layer from "effect/Layer";
 import * as Ref from "effect/Ref";
 
-export interface DesktopStateShape {
-  readonly backendReady: Ref.Ref<boolean>;
-  readonly quitting: Ref.Ref<boolean>;
-}
-
-export class DesktopState extends Context.Service<DesktopState, DesktopStateShape>()(
-  "@t3tools/desktop/app/DesktopState",
-) {}
-
-export const layer = Layer.effect(
+export class DesktopState extends Context.Service<
   DesktopState,
-  Effect.all({
-    backendReady: Ref.make(false),
-    quitting: Ref.make(false),
-  }),
-);
+  {
+    readonly backendReady: Ref.Ref<boolean>;
+    readonly quitting: Ref.Ref<boolean>;
+  }
+>()("@t3tools/desktop/app/DesktopState") {}
+
+const make = Effect.all({
+  backendReady: Ref.make(false),
+  quitting: Ref.make(false),
+});
+
+export const layer = Layer.effect(DesktopState, make);
